[IFT] Move glyph keyed functionality into it's own module.
